ml_model/requirements.txt
numpy
pandas
scikit-learn
tensorflow
fastapi
uvicorn
pydantic
joblib
python-dotenv
matplotlib
seaborn
python-dotenv
pymongo
dnspython